requirements.txt
google-generativeai>=0.3.0
streamlit>=1.28.0
Pillow>=9.0.0